Address the second round of review comments
- Split `WordBatchBuilder` and `VocabularyWriter` out of `MergePipeline.h`
  into `WordBatchBuilder.h` and `VocabularyWriter.h`, so that each stage of
  the merging pipeline stands on its own next to its existing test.
- Be precise about why `IdMapEntry::localIndex_` is a plain index: the cost
  is the range check in `Id::makeFromVocabIndex`, not the datatype bits.
- Log the start of a batch at `TRACE` instead of `TIMING` (both sites), as
  there is one such message per batch.
- Let `IdMapFromPartialIdMapFile` return a `HashMap<VocabIndex, Id>`, which
  removes the `Id::makeFromVocabIndex` per entry and resolves the
  corresponding TODO in `IndexImpl::convertPartialToGlobalIds`. The values
  stay `Id`s, because a merged word may also become a blank node.
